move haaqi test to the right module
